requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bleach==5.0.0
Django==4.0.4
django-bootstrap4==22.1
django-filter==21.1
django-jazzmin==2.5.0
django-summernote==0.8.20.0
django-taggit==3.0.0
Pillow==9.1.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
webencodings==0.5.1
